refactor(ecovent): split protocol client modules
