+++ custom_components/ecovent_v2/sensor.py
@@ -2,20 +2,16 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 import logging
 import re
 
 from .ecoventv2 import Fan
 
 from homeassistant.components.sensor import (
-    SensorDeviceClass,
     SensorEntity,
-    SensorStateClass,
 )
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import PERCENTAGE, REVOLUTIONS_PER_MINUTE, EntityCategory
-from homeassistant.const import UnitOfTemperature
+from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
@@ -27,432 +23,7 @@ from .coordinator import EcoVentCoordinator
 _LOGGER = logging.getLogger(__name__)
 
 
-@dataclass(frozen=True)
-class SensorSpec:
-    """Sensor description guarded by protocol parameters and capabilities."""
-
-    key: str
-    name: str
-    method: str
-    native_unit_of_measurement: str | None = None
-    device_class: SensorDeviceClass | None = None
-    state_class: SensorStateClass | None = None
-    entity_category: EntityCategory | None = None
-    enable_by_default: bool = True
-    icon: str | None = None
-    translation_key: str | None = None
-    suggested_display_precision: int | None = None
-    required_params: tuple[str, ...] = ()
-    required_capabilities: tuple[str, ...] = ()
-    excluded_params: tuple[str, ...] = ()
-    excluded_capabilities: tuple[str, ...] = ()
-
-
-SENSOR_SPECS = (
-    SensorSpec(
-        "_humidity",
-        "Humidity",
-        "humidity",
-        PERCENTAGE,
-        SensorDeviceClass.HUMIDITY,
-        SensorStateClass.MEASUREMENT,
-        icon="mdi:water-percent",
-    ),
-    SensorSpec(
-        "_temperature",
-        "Temperature",
-        "temperature",
-        UnitOfTemperature.CELSIUS,
-        SensorDeviceClass.TEMPERATURE,
-        SensorStateClass.MEASUREMENT,
-        icon="mdi:thermometer",
-        required_capabilities=("temperature",),
-    ),
-    SensorSpec(
-        "_room_temperature",
-        "Room temperature",
-        "room_temperature",
-        UnitOfTemperature.CELSIUS,
-        SensorDeviceClass.TEMPERATURE,
-        SensorStateClass.MEASUREMENT,
-        icon="mdi:home-thermometer-outline",
-        required_params=("room_temperature",),
-    ),
-    SensorSpec(
-        "_outdoor_temperature",
-        "Outdoor temperature",
-        "outdoor_temperature",
-        UnitOfTemperature.CELSIUS,
-        SensorDeviceClass.TEMPERATURE,
-        SensorStateClass.MEASUREMENT,
-        icon="mdi:thermometer-lines",
-        required_capabilities=("temperature_probes",),
-    ),
-    SensorSpec(
-        "_supply_temperature",
-        "Supply temperature",
-        "supply_temperature",
-        UnitOfTemperature.CELSIUS,
-        SensorDeviceClass.TEMPERATURE,
-        SensorStateClass.MEASUREMENT,
-        icon="mdi:thermometer-chevron-up",
-        required_capabilities=("temperature_probes",),
-    ),
-    SensorSpec(
-        "_exhaust_in_temperature",
-        "Exhaust inlet temperature",
-        "exhaust_in_temperature",
-        UnitOfTemperature.CELSIUS,
-        SensorDeviceClass.TEMPERATURE,
-        SensorStateClass.MEASUREMENT,
-        icon="mdi:thermometer-chevron-down",
-        required_capabilities=("temperature_probes",),
-    ),
-    SensorSpec(
-        "_exhaust_out_temperature",
-        "Exhaust outlet temperature",
-        "exhaust_out_temperature",
-        UnitOfTemperature.CELSIUS,
-        SensorDeviceClass.TEMPERATURE,
-        SensorStateClass.MEASUREMENT,
-        icon="mdi:thermometer-low",
-        required_capabilities=("temperature_probes",),
-    ),
-    SensorSpec(
-        "_co2",
-        "CO2",
-        "co2",
-        "ppm",
-        state_class=SensorStateClass.MEASUREMENT,
-        icon="mdi:molecule-co2",
-        required_capabilities=("co2",),
-    ),
-    SensorSpec(
-        "_voc",
-        "VOC",
-        "voc",
-        "index",
-        state_class=SensorStateClass.MEASUREMENT,
-        icon="mdi:air-filter",
-        required_capabilities=("voc",),
-    ),
-    SensorSpec(
-        "_air_quality",
-        "Air quality",
-        "air_quality",
-        "index",
-        state_class=SensorStateClass.MEASUREMENT,
-        icon="mdi:air-filter",
-        required_params=("air_quality",),
-    ),
-    SensorSpec(
-        "_speed",
-        "Fan speed",
-        "fan1_speed",
-        REVOLUTIONS_PER_MINUTE,
-        state_class=SensorStateClass.MEASUREMENT,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:fan",
-        excluded_params=("fan2_speed",),
-    ),
-    SensorSpec(
-        "_speed1",
-        "Fan 1 speed",
-        "fan1_speed",
-        REVOLUTIONS_PER_MINUTE,
-        state_class=SensorStateClass.MEASUREMENT,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:fan-speed-1",
-        required_params=("fan1_speed", "fan2_speed"),
-    ),
-    SensorSpec(
-        "_speed2",
-        "Fan 2 speed",
-        "fan2_speed",
-        REVOLUTIONS_PER_MINUTE,
-        state_class=SensorStateClass.MEASUREMENT,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        enable_by_default=False,
-        icon="mdi:fan-speed-2",
-        required_params=("fan2_speed",),
-    ),
-    SensorSpec(
-        "_airflow",
-        "Airflow",
-        "airflow",
-        device_class=SensorDeviceClass.ENUM,
-        icon="mdi:hvac",
-        translation_key="airflow",
-        required_capabilities=("airflow",),
-    ),
-    SensorSpec(
-        "_beeper",
-        "Beeper",
-        "beeper",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:volume-high",
-        translation_key="beeper",
-        required_capabilities=("beeper_status",),
-    ),
-    SensorSpec(
-        "_battery_status",
-        "Battery status",
-        "battery_status",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:battery",
-        translation_key="battery_status",
-        required_capabilities=("battery_status",),
-    ),
-    SensorSpec(
-        "_boost_status",
-        "Boost status",
-        "boost_status",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:fan-chevron-up",
-        translation_key="boost_status",
-        required_capabilities=("toggle_boost_status",),
-    ),
-    SensorSpec(
-        "_boost_status",
-        "Boost status",
-        "boost_status",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:fan-chevron-up",
-        translation_key="boost_status",
-        required_capabilities=("boost_delay_status",),
-    ),
-    SensorSpec(
-        "_timer_mode",
-        "Timer mode",
-        "timer_mode",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        enable_by_default=False,
-        icon="mdi:timer-cog-outline",
-        translation_key="timer_mode",
-        required_capabilities=("timer_mode",),
-    ),
-    SensorSpec(
-        "_timer_status",
-        "Timer status",
-        "timer_status",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        enable_by_default=False,
-        icon="mdi:timer-check-outline",
-        required_capabilities=("arc_environment",),
-    ),
-    SensorSpec(
-        "_alarm_status",
-        "Alarm status",
-        "alarm_status",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:alert-outline",
-        translation_key="alarm_status",
-    ),
-    SensorSpec(
-        "_alarm_list",
-        "Alarm list",
-        "alarm_list",
-        entity_category=EntityCategory.DIAGNOSTIC,
-        enable_by_default=False,
-        icon="mdi:alert-box-outline",
-        required_capabilities=("air_quality",),
-    ),
-    SensorSpec(
-        "_heater_status",
-        "Heater status",
-        "heater_status",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:heat-wave",
-        required_capabilities=("heater",),
-    ),
-    SensorSpec(
-        "_air_quality_status",
-        "Air quality status",
-        "air_quality_status",
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:air-filter",
-        required_capabilities=("air_quality",),
-    ),
-    SensorSpec(
-        "_air_quality_status",
-        "Air quality status",
-        "air_quality_status",
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:air-filter",
-        required_capabilities=("arc_environment",),
-    ),
-    SensorSpec(
-        "_recovery_efficiency",
-        "Recovery efficiency",
-        "recovery_efficiency",
-        PERCENTAGE,
-        state_class=SensorStateClass.MEASUREMENT,
-        icon="mdi:heat-pump-outline",
-        required_capabilities=("air_quality",),
-    ),
-    SensorSpec(
-        "_schedule_speed",
-        "Schedule speed",
-        "schedule_speed",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:calendar-clock",
-        required_capabilities=("air_quality",),
-    ),
-    SensorSpec(
-        "_frost_protection_status",
-        "Frost protection",
-        "frost_protection_status",
-        device_class=SensorDeviceClass.ENUM,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:snowflake-alert",
-        required_capabilities=("air_quality",),
-    ),
-    SensorSpec(
-        "_timer_counter",
-        "Timer counter",
-        "timer_counter",
-        "h",
-        SensorDeviceClass.DURATION,
-        SensorStateClass.MEASUREMENT,
-        EntityCategory.DIAGNOSTIC,
-        False,
-        "mdi:timer-play-outline",
-    ),
-    SensorSpec(
-        "_battery",
-        "Battery",
-        "battery_voltage",
-        PERCENTAGE,
-        SensorDeviceClass.BATTERY,
-        SensorStateClass.MEASUREMENT,
-        EntityCategory.DIAGNOSTIC,
-        icon="mdi:battery",
-        required_capabilities=("battery_voltage",),
-    ),
-    SensorSpec(
-        "_filter_change_in",
-        "Filter change in",
-        "filter_timer_countdown",
-        "h",
-        SensorDeviceClass.DURATION,
-        SensorStateClass.MEASUREMENT,
-        EntityCategory.DIAGNOSTIC,
-        icon="mdi:timer-sand",
-        suggested_display_precision=1,
-        required_capabilities=("filter_maintenance",),
-    ),
-    SensorSpec(
-        "_filter_remaining",
-        "Filter remaining",
-        "filter_remaining",
-        PERCENTAGE,
-        state_class=SensorStateClass.MEASUREMENT,
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:air-filter",
-        suggested_display_precision=0,
-        required_params=("filter_timer_countdown", "filter_timer_setpoint"),
-        required_capabilities=("filter_maintenance",),
-    ),
-    SensorSpec(
-        "_night_mode_timer",
-        "Night mode timer",
-        "night_mode_timer",
-        "h",
-        SensorDeviceClass.DURATION,
-        SensorStateClass.MEASUREMENT,
-        EntityCategory.DIAGNOSTIC,
-        False,
-        "mdi:weather-night",
-        suggested_display_precision=2,
-        required_capabilities=("night_party_timers",),
-    ),
-    SensorSpec(
-        "_party_mode_timer",
-        "Party mode timer",
-        "party_mode_timer",
-        "h",
-        SensorDeviceClass.DURATION,
-        SensorStateClass.MEASUREMENT,
-        EntityCategory.DIAGNOSTIC,
-        False,
-        "mdi:party-popper",
-        suggested_display_precision=2,
-        required_capabilities=("night_party_timers",),
-    ),
-    SensorSpec(
-        "_analogv",
-        "Analog voltage",
-        "analogv",
-        entity_category=EntityCategory.DIAGNOSTIC,
-        enable_by_default=False,
-        icon="mdi:flash",
-        required_params=("analogV",),
-        required_capabilities=("analog_voltage",),
-    ),
-    SensorSpec(
-        "_machine_hours",
-        "Machine hours",
-        "machine_hours",
-        "h",
-        SensorDeviceClass.DURATION,
-        SensorStateClass.MEASUREMENT,
-        EntityCategory.DIAGNOSTIC,
-        False,
-        "mdi:timer-outline",
-    ),
-    SensorSpec(
-        "_screen_off_start_time",
-        "Screen off start time",
-        "screen_off_start_time",
-        entity_category=EntityCategory.DIAGNOSTIC,
-        enable_by_default=False,
-        icon="mdi:clock-start",
-        required_capabilities=("breezy_screen",),
-    ),
-    SensorSpec(
-        "_screen_off_end_time",
-        "Screen off end time",
-        "screen_off_end_time",
-        entity_category=EntityCategory.DIAGNOSTIC,
-        enable_by_default=False,
-        icon="mdi:clock-end",
-        required_capabilities=("breezy_screen",),
-    ),
-    SensorSpec(
-        "_silent_mode_start_time",
-        "Silent mode start time",
-        "silent_mode_start_time",
-        entity_category=EntityCategory.DIAGNOSTIC,
-        enable_by_default=False,
-        icon="mdi:clock-start",
-        required_capabilities=("arc_environment",),
-    ),
-    SensorSpec(
-        "_silent_mode_end_time",
-        "Silent mode end time",
-        "silent_mode_end_time",
-        entity_category=EntityCategory.DIAGNOSTIC,
-        enable_by_default=False,
-        icon="mdi:clock-end",
-        required_capabilities=("arc_environment",),
-    ),
-    SensorSpec(
-        "_ip",
-        "IP address",
-        "current_wifi_ip",
-        entity_category=EntityCategory.DIAGNOSTIC,
-        icon="mdi:ip-network",
-    ),
-)
+from .sensor_specs import SENSOR_SPECS
 
 
 def _parse_hours(
